(refactor): clean up clutter
interfaces
tests
services
