+++ src/app/ng-aws-cognito/index.ts
@@ -1,5 +1,5 @@
 export { NgAwsCognitoModule } from "./ng-aws-cognito.module";
-export { UserService } from "./user.service";
-export { LocalStorageService } from "./local-storage.service";
-export { AwsService } from "./aws.service";
-export { ApiClientService } from "./api-client.service";
+export { UserService } from "./services/user.service";
+export { LocalStorageService } from "./services/local-storage.service";
+export { AwsService } from "./services/aws.service";
+export { ApiClientService } from "./services/api-client.service";
