+++ src/app/ng-aws-cognito/interfaces/aws-cognito-config.ts
