+++ src/app/ng-aws-cognito/interfaces/iam-credentials.ts
