+++ src/app/ng-aws-cognito/ng-aws-cognito.module.ts
@@ -1,18 +1,18 @@
 import { NgModule, ModuleWithProviders, InjectionToken } from "@angular/core";
 import { CommonModule } from '@angular/common';
 
-import { ApiClientService } from "./api-client.service";
-import { ApiGatewayService } from "./api-gateway.service";
-import { AwsService } from "./aws.service";
+import { ApiClientService } from "./services/api-client.service";
+import { ApiGatewayService } from "./services/api-gateway.service";
+import { AwsService } from "./services/aws.service";
 import { COGNITO_CONFIG } from "./cognito-config-token";
 import { ApigClientFactory } from "./apig-client-factory";
-import { UserService } from './user.service';
-import { LocalStorageService } from './local-storage.service';
+import { UserService } from './services/user.service';
+import { LocalStorageService } from './services/local-storage.service';
 
 
 @NgModule({
     imports: [
-        CommonModule 
+        CommonModule
     ],
     providers: [
         ApiClientService,
@@ -39,13 +39,13 @@ export class NgAwsCognitoModule {
 //import { NgModule, ModuleWithProviders, InjectionToken } from "@angular/core";
 //import { CommonModule } from "@angular/common";
 
-/*import { ApiClientService } from "./api-client.service";
-import { ApiGatewayService } from "./api-gateway.service";
-import { AwsService } from "./aws.service";
+/*import { ApiClientService } from "./api-client.services";
+import { ApiGatewayService } from "./api-gateway.services";
+import { AwsService } from "./aws.services";
 import { COGNITO_CONFIG } from "./cognito-config-token";
 import { ApigClientFactory } from "./apig-client-factory";
-import { UserService } from './user.service';
-import { LocalStorageService } from './local-storage.service';
+import { UserService } from './user.services';
+import { LocalStorageService } from './local-storage.services';
 
 @NgModule({
