+++ src/app/ng-aws-cognito/services/local-storage.service.ts
