+++ src/app/ng-aws-cognito/tests/local-storage.service.spec.ts
@@ -2,7 +2,7 @@
 
 import { TestBed, async, inject } from "@angular/core/testing";
 
-import { LocalStorageService } from "../local-storage.service";
+import { LocalStorageService } from "../services/local-storage.service";
 
 describe("Service: LocalStorageService", () => {
     beforeEach(() => {
