+++ src/app/ng2-aws-cognito.spec.ts
@@ -1,5 +1,7 @@
-/*require('../../../tests/api-client.service.spec');
+/*
+require('../../../tests/api-client.service.spec');
 require('../../../tests/api-gateway.service.spec');
 require('../../../tests/aws.service.spec');
 require('../../../tests/local-storage.service.spec');
-require('../../../tests/user.service.spec');*/
+require('../../../tests/user.service.spec');
+*/
